Conflicts:
	lib/Cake/View/Helper/HtmlHelper.php
	lib/Cake/View/Helper/PaginatorHelper.php

+++ lib/Cake/Cache/Engine/FileEngine.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * File Storage engine for cache.  Filestorage is the slowest cache storage
- * to read and write.  However, it is good for servers that don't have other storage
+ * File Storage engine for cache. Filestorage is the slowest cache storage
+ * to read and write. However, it is good for servers that don't have other storage
  * engine available, or have content which is not performance sensitive.
  *
  * You can configure a FileEngine cache, using Cache::config()
@@ -9,20 +9,21 @@
  * PHP 5
  *
  * CakePHP(tm) : Rapid Development Framework (http://cakephp.org)
- * Copyright 2005-2012, Cake Software Foundation, Inc. (http://cakefoundation.org)
+ * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
  *
  * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
  * Redistributions of files must retain the above copyright notice.
  *
- * @copyright     Copyright 2005-2012, Cake Software Foundation, Inc. (http://cakefoundation.org)
+ * @copyright     Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
  * @link          http://cakephp.org CakePHP(tm) Project
  * @since         CakePHP(tm) v 1.2.0.4933
- * @license       MIT License (http://www.opensource.org/licenses/mit-license.php)
+ * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
 
 /**
- * File Storage engine for cache.  Filestorage is the slowest cache storage
- * to read and write.  However, it is good for servers that don't have other storage
+ * File Storage engine for cache. Filestorage is the slowest cache storage
+ * to read and write. However, it is good for servers that don't have other storage
  * engine available, or have content which is not performance sensitive.
  *
  * You can configure a FileEngine cache, using Cache::config()
@@ -94,7 +95,7 @@ class FileEngine extends CacheEngine {
 /**
  * Garbage collection. Permanently remove all expired and deleted data
  *
- * @param integer $expires [optional] An expires timestamp, invalidataing all data before.
+ * @param integer $expires [optional] An expires timestamp, invalidating all data before.
  * @return boolean True if garbage collection was successful, false on failure
  */
 	public function gc($expires = null) {
@@ -358,16 +359,18 @@ class FileEngine extends CacheEngine {
  * Recursively deletes all files under any directory named as $group
  *
  * @return boolean success
- **/
+ */
 	public function clearGroup($group) {
 		$directoryIterator = new RecursiveDirectoryIterator($this->settings['path']);
 		$contents = new RecursiveIteratorIterator($directoryIterator, RecursiveIteratorIterator::CHILD_FIRST);
 		foreach ($contents as $object) {
 			$containsGroup = strpos($object->getPathName(), DS . $group . DS) !== false;
-			if ($object->isFile() && $containsGroup) {
+			$hasPrefix = strpos($object->getBaseName(), $this->settings['prefix']) === 0;
+			if ($object->isFile() && $containsGroup && $hasPrefix) {
 				unlink($object->getPathName());
 			}
 		}
+		$this->_File = null;
 		return true;
 	}
 }
